+++ example/actions/chat.ts
@@ -1,11 +1,15 @@
 import { z } from "zod";
+import { Metadata } from "../config";
 import { actionsFactory } from "../factories";
 
 export const onChat = actionsFactory.build({
   input: z.tuple([z.string()]),
-  handler: async ({ input: [message], broadcast, logger }) => {
+  handler: async ({ input: [message], client, all, logger }) => {
     try {
-      broadcast("chat", message);
+      await all.broadcast("chat", message, { from: client.id });
+      client.setData<Metadata>({
+        msgCount: (client.getData<Metadata>().msgCount || 0) + 1,
+      });
     } catch (error) {
       logger.error("Failed to broadcast", error);
     }