+++ example/config.ts
@@ -1,14 +1,21 @@
 import { z } from "zod";
-import { createSocketsConfig } from "../src";
+import { createConfig } from "../src";
 
-export const socketsConfig = createSocketsConfig({
+/** @desc Client metadata */
+export interface Metadata {
+  /** @desc Number of messages sent using the chat event */
+  msgCount: number;
+}
+
+export const config = createConfig({
   timeout: 2000,
+  logger: console,
   emission: {
     time: {
       schema: z.tuple([z.date().transform((date) => date.toISOString())]),
     },
     chat: {
-      schema: z.tuple([z.string()]),
+      schema: z.tuple([z.string(), z.object({ from: z.string() })]),
     },
   },
 });