+++ src/actions-factory.ts
@@ -1,31 +1,36 @@
 import { z } from "zod";
 import { Action, Handler } from "./action";
+import { Config } from "./config";
 import { EmissionMap } from "./emission";
-import { SocketsConfig } from "./index";
 
-export interface SimpleActionDef<
+export interface ActionNoAckDef<
   IN extends z.AnyZodTuple,
   E extends EmissionMap,
 > {
+  /** @desc The incoming event payload validation schema (no acknowledgement) */
   input: IN;
+  /** @desc No output schema => no returns => no acknowledgement */
   handler: Handler<z.output<IN>, void, E>;
 }
 
-export interface AckActionDef<
+export interface ActionWithAckDef<
   IN extends z.AnyZodTuple,
   OUT extends z.AnyZodTuple,
   E extends EmissionMap,
 > {
+  /** @desc The incoming event payload (excl. acknowledgement) validation schema */
   input: IN;
+  /** @desc The acknowledgement validation schema */
   output: OUT;
+  /** @desc The returns become an Acknowledgement */
   handler: Handler<z.output<IN>, z.input<OUT>, E>;
 }
 
 export class ActionsFactory<E extends EmissionMap> {
-  constructor(protected config: SocketsConfig<E>) {}
+  constructor(protected config: Config<E>) {}
 
   public build<IN extends z.AnyZodTuple, OUT extends z.AnyZodTuple>(
-    def: SimpleActionDef<IN, E> | AckActionDef<IN, OUT, E>,
+    def: ActionNoAckDef<IN, E> | ActionWithAckDef<IN, OUT, E>,
   ): Action<IN, OUT> {
     return new Action(def);
   }