+++ src/emission.ts
@@ -1,6 +1,8 @@
+import assert from "node:assert/strict";
 import type { Socket } from "socket.io";
 import { z } from "zod";
-import { AbstractLogger } from "./logger";
+import { Config } from "./config";
+import { RemoteClient, getRemoteClients } from "./remote-client";
 
 export interface Emission {
   schema: z.AnyZodTuple;
@@ -12,37 +14,92 @@ export interface EmissionMap {
 }
 
 type TupleOrTrue<T> = T extends z.AnyZodTuple ? T : z.ZodLiteral<true>;
+type TuplesOrTrue<T> = T extends z.AnyZodTuple
+  ? z.ZodArray<T>
+  : z.ZodLiteral<true>;
 
 export type Emitter<E extends EmissionMap> = <K extends keyof E>(
   evt: K,
   ...args: z.input<E[K]["schema"]>
 ) => Promise<z.output<TupleOrTrue<E[K]["ack"]>>>;
 
+export type Broadcaster<E extends EmissionMap> = <K extends keyof E>(
+  evt: K,
+  ...args: z.input<E[K]["schema"]>
+) => Promise<z.output<TuplesOrTrue<E[K]["ack"]>>>;
+
+export type RoomService<E extends EmissionMap> = (rooms: string | string[]) => {
+  /**
+   * @desc Emits an event to everyone in the specified room(s)
+   * @throws z.ZodError on validation
+   * @throws Error on ack timeout
+   * */
+  broadcast: Broadcaster<E>;
+  join: () => void | Promise<void>;
+  leave: () => void | Promise<void>;
+  getClients: () => Promise<RemoteClient[]>;
+};
+
 /**
  * @throws z.ZodError on validation
  * @throws Error on ack timeout
  * */
-export const makeEmitter =
-  <E extends EmissionMap>({
-    emission,
-    logger,
-    socket,
-    timeout,
+const makeGenericEmitter =
+  ({
+    target,
+    config: { logger, emission, timeout },
   }: {
-    emission: E;
-    logger: AbstractLogger;
-    socket: Socket;
-    timeout: number;
-  }): Emitter<E> =>
-  async (event, ...args) => {
-    const { schema, ack: ackSchema } = emission[event];
+    config: Config<EmissionMap>;
+    target: Socket | Socket["broadcast"];
+  }) =>
+  async (event: string, ...args: unknown[]) => {
+    const isSocket = "id" in target;
+    assert(event in emission, new Error(`Unsupported event ${event}`));
+    const { schema, ack } = emission[event];
     const payload = schema.parse(args);
-    logger.debug(`Emitting ${String(event)}`, payload);
-    if (!ackSchema) {
-      return socket.emit(String(event), ...payload) || true;
+    logger.debug(
+      `${isSocket ? "Emitting" : "Broadcasting"} ${String(event)}`,
+      payload,
+    );
+    if (!ack) {
+      return target.emit(String(event), ...payload) || true;
     }
-    const ack = await socket
+    const response = await target
       .timeout(timeout)
       .emitWithAck(String(event), ...payload);
-    return ackSchema.parse(ack);
+    return (isSocket ? ack : ack.array()).parse(response);
   };
+
+interface MakerParams<E extends EmissionMap> {
+  socket: Socket;
+  config: Config<E>;
+}
+
+export const makeEmitter = <E extends EmissionMap>({
+  socket: target,
+  ...rest
+}: MakerParams<E>) => makeGenericEmitter({ ...rest, target }) as Emitter<E>;
+
+export const makeBroadcaster = <E extends EmissionMap>({
+  socket: { broadcast: target },
+  ...rest
+}: MakerParams<E>) => makeGenericEmitter({ ...rest, target }) as Broadcaster<E>;
+
+export const makeRoomService =
+  <E extends EmissionMap>({
+    socket,
+    ...rest
+  }: MakerParams<E>): RoomService<E> =>
+  (rooms) => ({
+    getClients: async () =>
+      getRemoteClients(await socket.in(rooms).fetchSockets()),
+    join: () => socket.join(rooms),
+    leave: () =>
+      typeof rooms === "string"
+        ? socket.leave(rooms)
+        : Promise.all(rooms.map((room) => socket.leave(room))).then(() => {}),
+    broadcast: makeGenericEmitter({
+      ...rest,
+      target: socket.to(rooms),
+    }) as Broadcaster<E>,
+  });