+++ src/index.ts
@@ -1,59 +1,8 @@
-import http from "node:http";
-import type { Server } from "socket.io";
-import { ActionMap, Handler, SocketFeatures } from "./action";
-import { makeBroadcaster } from "./broadcasting";
-import { EmissionMap, makeEmitter } from "./emission";
-import { AbstractLogger } from "./logger";
+export { attachSockets } from "./attach";
+export { createConfig, type Config } from "./config";
+export { ActionsFactory } from "./actions-factory";
+export type { ActionMap, HandlingFeatures } from "./action";
 
-export interface SocketsConfig<E extends EmissionMap> {
-  timeout: number;
-  emission: E;
-}
-
-export const createSocketsConfig = <E extends EmissionMap>(
-  config: SocketsConfig<E>,
-) => config;
-
-export const attachSockets = <E extends EmissionMap>({
-  io,
-  actions,
-  logger,
-  target,
-  config: { emission, timeout },
-  onConnection = ({ socketId }) => logger.debug("User connected", socketId),
-  onDisconnect = ({ socketId }) => logger.debug("User disconnected", socketId),
-  onAnyEvent = ({ input: [event], socketId }) =>
-    logger.debug(`${event} from ${socketId}`),
-}: {
-  io: Server;
-  actions: ActionMap;
-  logger: AbstractLogger;
-  target: http.Server;
-  config: SocketsConfig<E>;
-  onConnection?: Handler<[], void, E>;
-  onDisconnect?: Handler<[], void, E>;
-  onAnyEvent?: Handler<[string], void, E>;
-}): Server => {
-  logger.info("ZOD-SOCKETS", target.address());
-  io.on("connection", async (socket) => {
-    const commons: SocketFeatures = {
-      socketId: socket.id,
-      isConnected: () => socket.connected,
-    };
-    const emit = makeEmitter({ emission, socket, logger, timeout });
-    const broadcast = makeBroadcaster({ emission, socket, logger, timeout });
-    await onConnection({ input: [], logger, emit, broadcast, ...commons });
-    socket.onAny((event) =>
-      onAnyEvent({ input: [event], logger, emit, broadcast, ...commons }),
-    );
-    for (const [event, action] of Object.entries(actions)) {
-      socket.on(event, async (...params) =>
-        action.execute({ event, params, logger, emit, broadcast, ...commons }),
-      );
-    }
-    socket.on("disconnect", () =>
-      onDisconnect({ input: [], logger, emit, broadcast, ...commons }),
-    );
-  });
-  return io.attach(target);
-};
+// issue 952
+export type { EmissionMap } from "./emission";
+export type { LoggerOverrides } from "./logger";