+++ src/remote-client.ts
@@ -1,0 +1,14 @@
+import { RemoteSocket } from "socket.io";
+
+export interface RemoteClient {
+  id: string;
+  rooms: string[];
+  getData: <D extends object>() => Readonly<Partial<D>>;
+}
+
+export const getRemoteClients = (sockets: RemoteSocket<{}, unknown>[]) =>
+  sockets.map<RemoteClient>(({ id, rooms, data }) => ({
+    id: id,
+    rooms: Array.from(rooms),
+    getData: <D extends object>() => (data as D) || {},
+  }));