+++ tests/cjs/quick-start.spec.ts
@@ -1,0 +1,33 @@
+import { spawn } from "node:child_process";
+import { afterAll, afterEach, describe, test } from "vitest";
+import { waitFor } from "../helpers";
+
+describe("CJS Test", async () => {
+  let out = "";
+  const listener = (chunk: Buffer) => {
+    out += chunk.toString();
+  };
+  const quickStart = spawn("tsx", ["quick-start.ts"], {
+    cwd: "./tests/cjs",
+  });
+  quickStart.stdout.on("data", listener);
+  quickStart.stderr.on("data", listener);
+
+  afterAll(async () => {
+    quickStart.stdout.removeListener("data", listener);
+    quickStart.stderr.removeListener("data", listener);
+    quickStart.kill();
+    await waitFor(() => quickStart.killed);
+  });
+
+  afterEach(() => {
+    console.log(out);
+    out = "";
+  });
+
+  describe("Quick Start from Readme", () => {
+    test("should start", async () => {
+      await waitFor(() => out.indexOf("ZOD-SOCKETS") > -1);
+    });
+  });
+});