+++ vitest.config.ts
@@ -1,0 +1,13 @@
+import { defineConfig } from "vitest/config";
+
+export default defineConfig({
+  test: {
+    testTimeout: 10000,
+    reporters: "verbose",
+    coverage: {
+      provider: "istanbul",
+      reporter: [["text", { maxCols: 120 }], "json-summary", "html", "lcov"],
+      include: ["src/**"],
+    },
+  },
+});